feat: add global config var for w3id
